feat: Phase 185 Tier3 auto-control enablement + Space tab rendering fix
Tier3 Auto-Control Enablement (Phase 185, Waves 1+3):
- app/models/sentinel_tool.py: SentinelTool dataclass (6-category classification, dangerous-gate demotion, tool_metadata audit payload)
- app/models/onboarding_phase.py: check_shadow_exit_criteria() (5 gates for Advisory→Supervised flip)
- app/services/chat_tools.py: safety_profiles matrix for control_device action type
- app/services/tier_routing_engine.py: LLM confidence bounding (ML_TRUST_WEIGHT_MIN×0.95)
- app/services/approval_service.py: audit_level + context_snapshot capture in parasite_decisions
- app/services/system_health_service.py: rename probes (niagara→supervisor, bacnet→field_network); OBIX client import fixed; lighting .count("exact") chaining bug fixed
- app/api/system_health.py: bms_connectivity aggregate (supervisor/field_network/obix/lighting)
- backend/app/database/repositories/security_repository.py: get_occupancy() rewritten for access_events table
- app/services/shadow_mode_polling.py: S002-CHILLER-AGG/S002-SITE-AGG → site_aggregate; occupancy wired
- app/services/sentinel_ml_feeder.py: site_aggregate in SENSOR_MAPPING; derived features (hvac_ratio, lighting_ratio, non_hvac_kw)
- frontend cockpit: CockpitBuildingThree restored; isAutomatic typo fixed
- backend/tests/services/: test_ml_pipeline_integrity.py, test_mqtt_ingestion_flow.py, test_sentinel_ml_feeder.py

Space/Concierge Intelligence tab rendering fix:
- ConciergeDashboardPage.tsx: GSAP null guard + try-catch (map stayed invisible if ref was null)
- ConciergeMap.tsx: containerReady useState init fix; ResizeObserver for late DOM attachment; zero-dimension guard before cytoscape init

Documentation consolidation:
- NAMING_CONVENTIONS.md deleted from root (canonical: docs/02-architecture/NAMING_CONVENTIONS.md)
- CLAUDE_NAMING.md deleted (duplicate of canonical)
- 12 cross-reference fixes for naming-conventions.md → NAMING_CONVENTIONS.md
- docs/_archive/: CLAUDE_md_analysis.md, TODOdone2.md archived via git mv
- AGENTS.md + DOCUMENTATION_RULES.md: NAMING_CONVENTIONS.md removed from root exception list

Backend data cleanup (JSON → Supabase migration):
- Deleted stale data files: alerts.json, anomalies.json, desks.json, equipment.json, hvac_zones.json, predictions.json, readings.json, sensors.json, sites.json

+++ backend/app/services/lighting_simulation_engine.py
@@ -360,7 +360,7 @@ class LightingSimulationEngine:
             # Determine dim level (0-254 DALI scale)
             dim_ratio = per_lum_power_w / sample_rated if sample_rated > 0 else 0.0
             dim_ratio = min(1.0, dim_ratio)
-            dali_level = int(round(dim_ratio * 254))
+            dali_level = round(dim_ratio * 254)
 
             # Assign zone to a sceneCOM controller (1 per floor)
             zone_config = self._zone_cache.get(zone_id, {})
@@ -491,7 +491,7 @@ class LightingSimulationEngine:
             # --- Sensor telemetry (one per zone) ---
             # Derive occupancy count from percent and typical occupancy
             typical = zone_config.get("typical_occupancy", 10)
-            occ_count = max(0, int(round(occupancy_pct / 100.0 * typical)))
+            occ_count = max(0, round(occupancy_pct / 100.0 * typical))
             sensor_records.append(
                 {
                     "sensor_id": f"PIR-{zone_id}",
